Conflicts:
	source/ivec4.h
	source/main.cpp
	source/uvec4.h
	source/vec4.h

+++ source/ivec4.h
@@ -123,7 +123,7 @@ class ivec4
 			return v0;
 		}
 /*
-		friend inline ivec4& operator /= (ivec4 &v, float f) {
+		friend inline ivec4& operator /= (ivec4 &v, int32_t f) {
 			???
 		}
 
@@ -203,7 +203,7 @@ class ivec4
 			return _mm_xor_si128(_mm_add_epi32(v.m, mask), mask);
 		}
 
-		friend inline const ivec4 clamp(const ivec4 &v, float f1, float f2) {
+		friend inline const ivec4 clamp(const ivec4 &v, int32_t f1, int32_t f2) {
 			return max(min(v, f2), f1);
 		}
